Moved portfolio to container fro separation of states and non stateful components
