+++ src/components/FooterContent/index.css
@@ -4,7 +4,7 @@
   bottom: 0;
   width: 100%;
   height: 60px;
-
+  border-top: 3px solid orange;
   text-align: center;
   background-color: #24273f;
   color: #ffffff;
