+++ src/components/Navigation/index.css
@@ -1,0 +1,3 @@
+.navbar {
+  border-bottom: 3px solid orange;
+}
