+++ src/index.css
@@ -1,3 +1,11 @@
+html,
+body {
+  height: 100%;
+  background-color: #181a29;
+}
+#root {
+  background-color: #181a29;
+}
 a li {
   margin-top: 5px;
 }
@@ -9,29 +17,8 @@ a li:hover {
   min-height: 100%;
   padding-bottom: 60px;
 }
-.card {
-  width: 65%;
-  margin: auto;
-  display: block;
-  text-align: center;
-  border-bottom: 2px solid darkorange;
-}
-.card a {
-  color: darkorange;
-}
 .content a {
   margin-top: 10px;
-}
-.navbar {
-  border-bottom: 3px solid orange;
-}
-.card-header-title,
-.card-footer-item {
-  background-color: #24273f0a;
-  text-align: center;
-}
-.footer {
-  border-top: 3px solid orange;
 }
 @media screen and (max-width: 900px) {
   .card {
